refactor: move test doubles and fixtures to conftest
