+++ tests/unit/stream_processor/test_process_stream_handler.py
@@ -1,48 +1,26 @@
-from typing import Any
-
-from product.models.products.product import ProductNotification
-from product.stream_processor.dal.events.base import EventHandler
 from product.stream_processor.handlers.process_stream import process_stream
-from tests.unit.stream_processor.data_builder import generate_dynamodb_stream_events
+from tests.unit.stream_processor.conftest import FakeEventHandler
 from tests.utils import generate_context
 
 
-class FakeEventHandler(EventHandler):
-
-    def __init__(self):
-        self.published_events = []
-
-    def emit(self, payload: list[ProductNotification], metadata: dict[str, Any] | None = None):
-        metadata = metadata or {}
-        for product in payload:
-            self.published_events.append({'event': product, 'metadata': metadata})
-
-    def __len__(self):
-        return len(self.published_events)
-
-
-def test_process_stream_notify_product_updates():
+def test_process_stream_notify_product_updates(dynamodb_stream_events: dict, event_store: FakeEventHandler):
     # GIVEN a DynamoDB stream event and a fake event handler
-    event = generate_dynamodb_stream_events()
-    event_store = FakeEventHandler()
-
     # WHEN process_stream is called with a custom event handler
-    process_stream(event=event, context=generate_context(), event_handler=event_store)
+    process_stream(event=dynamodb_stream_events, context=generate_context(), event_handler=event_store)
 
-    # THEN the fake event handler should have received the correct number of events
+    # THEN the fake event handler should emit these product notifications
     # and no errors should have been raised (e.g., no sockets, no DAL calls)
-    assert len(event['Records']) == len(event_store)
+    assert len(dynamodb_stream_events['Records']) == len(event_store)
 
 
 # NOTE: this should fail once we have schema validation
-def test_process_stream_with_empty_records():
+def test_process_stream_with_empty_records(event_store: FakeEventHandler):
     # GIVEN an empty DynamoDB stream event
-    event = {'Records': []}
-    event_store = FakeEventHandler()
+    event: dict[str, list] = {'Records': []}
 
     # WHEN process_stream is called with a custom event handler
     process_stream(event=event, context=generate_context(), event_handler=event_store)
 
-    # THEN the fake event handler should have received no events
+    # THEN the fake event handler should emit these product notifications
     # and no errors should have been raised
     assert len(event_store) == 0
